Ensure that VerifyContinualTests is always called if continual tests started
